(Chore) Update Jest to v27 (#1728)
Updates jest to v27
