+++ package.json
@@ -108,14 +108,14 @@
     "@redux-saga/is": "^1.1.2",
     "@svgr/webpack": "5.5.0",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.5",
+    "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "^7.0.1",
     "@testing-library/user-event": "^13.1.9",
     "@types/enzyme": "^3.10.8",
     "@types/faker": "^5.5.6",
     "@types/invariant": "^2.2.34",
-    "@types/jest": "^26.0.21",
-    "@types/jsdom": "^16.2.7",
+    "@types/jest": "^26.0.24",
+    "@types/jsdom": "^16.2.13",
     "@types/leaflet": "^1.7.0",
     "@types/leaflet.markercluster": "^1.4.4",
     "@types/lodash": "^4.14.168",
@@ -136,8 +136,7 @@
     "@typescript-eslint/parser": "^4.23.0",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.0",
     "add-asset-html-webpack-plugin": "^3.2.0",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "^26.6.3",
+    "babel-jest": "^27.0.6",
     "babel-loader": "^8.2.2",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-inline-react-svg": "^2.0.1",
@@ -185,8 +184,8 @@
     "image-webpack-loader": "^7.0.1",
     "imports-loader": "^2.0.0",
     "ip": "1.1.5",
-    "jest": "^26.6.3",
-    "jest-cli": "^26.6.3",
+    "jest": "^27.0.6",
+    "jest-cli": "^27.0.6",
     "jest-fetch-mock": "^3.0.3",
     "jest-localstorage-mock": "^2.4.8",
     "jest-styled-components": "^7.0.3",
@@ -214,7 +213,7 @@
     "shelljs": "^0.8.4",
     "style-loader": "^2.0.0",
     "svg-url-loader": "^7.1.1",
-    "ts-jest": "^26.5.4",
+    "ts-jest": "^27.0.3",
     "ts-loader": "^9.2.3",
     "typescript": "^4.3.5",
     "typescript-plugin-styled-components": "^1.6.0",
