+++ src/signals/incident/components/IncidentPreview/components/MapSelect/index.test.js
@@ -1,76 +1,28 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright (C)  - 2021 Gemeente Amsterdam
-import { render, screen } from '@testing-library/react'
-
 import configuration from 'shared/services/configuration/configuration'
+import { render, screen } from '@testing-library/react'
 import incidentFixture from 'utils/__tests__/fixtures/incident.json'
 import { withAppContext } from 'test/utils'
 
 jest.mock('shared/services/configuration/configuration')
 
-describe('signals/incident/components/IncidentPreview/components/MapSelect', () => {
-  afterEach(() => {
-    configuration.__reset()
-  })
-
-  it('renders correctly', () => {
-    jest.isolateModules(() => {
-      const MapSelect = require('.').default
-      const value = ['foo', 'bar', 'baz']
-
-      render(
-        withAppContext(
-          <MapSelect
-            value={value}
-            meta={{ endpoint: 'https://endpoint', idField: '' }}
-            incident={incidentFixture}
-          />
-        )
-      )
-
-      expect(screen.getByText(value.join('; '))).toBeInTheDocument()
-      expect(screen.getByTestId('mapSelect')).toBeInTheDocument()
-    })
-  })
-
+describe('signals/incident/components/IncidentPreview/components/MapSelect/index.js', () => {
   it('should render MapSelectGeneric with feature flag enabled', () => {
-    jest.isolateModules(() => {
-      configuration.featureFlags.useMapSelectGeneric = true
-      const MapSelect = require('.').default
-      const value = ['foo', 'bar', 'baz']
+    configuration.featureFlags.useMapSelectGeneric = true
+    const MapSelect = require('./index.js').default
+    const value = ['foo', 'bar', 'baz']
 
-      render(
-        withAppContext(
-          <MapSelect
-            value={value}
-            meta={{ endpoint: 'https://endpoint', idField: '' }}
-            incident={incidentFixture}
-          />
-        )
+    render(
+      withAppContext(
+        <MapSelect
+          value={value}
+          meta={{ endpoint: 'https://endpoint', idField: '' }}
+          incident={incidentFixture}
+        />
       )
+    )
 
-      expect(screen.getByTestId('mapSelectGeneric')).toBeInTheDocument()
-    })
-  })
-
-  it('returns a location', () => {
-    jest.isolateModules(() => {
-      const getLatlng = require('.').getLatlng
-      expect(getLatlng(incidentFixture.location)).toEqual({
-        latitude: incidentFixture.location.geometrie.coordinates[1],
-        longitude: incidentFixture.location.geometrie.coordinates[0],
-      })
-    })
-  })
-
-  it('returns default coords', () => {
-    jest.isolateModules(() => {
-      const getLatlng = require('.').getLatlng
-      const DEFAULT_COORDS = require('.').DEFAULT_COORDS
-      expect(getLatlng()).toEqual({
-        latitude: DEFAULT_COORDS[1],
-        longitude: DEFAULT_COORDS[0],
-      })
-    })
+    expect(screen.getByTestId('mapSelectGeneric')).toBeInTheDocument()
   })
 })
